+++ client/app/integrations/serialCom/actions/message/message.js
@@ -1,12 +1,33 @@
 // @flow
 
 import type { ActionType } from '../../../';
+import type { optionsValuesType, valuesEntries } from '../../../base';
+import TriggerOrAction from '../../../base';
 
-class Message implements ActionType {
-  static options = [];
+class Message extends TriggerOrAction implements ActionType {
+  static availableChars = (): valuesEntries => {
+    return [{ value: 'a', label: 'A' }];
+  };
+
+  static options = [
+    {
+      ID: 'payload',
+      required: true,
+      values: Message.availableChars,
+      controlType: 'select'
+    }
+  ];
+
   static metadata = {
     name: 'Serial port message',
     description: 'Send a message evia the serial port',
+  }
+
+  payload: string;
+
+  constructor(optionValues: optionsValuesType) {
+    super(Message.options, optionValues);
+    this.payload = optionValues.payload;
   }
 }
